+++ torch_timeseries/model/__init__.py
@@ -19,6 +19,7 @@ from .TimeDiff import TimeDiff
 from .NsDiff import NsDiff
 from .TMDM import TMDM
 from .SegRNN import SegRNN
+from .TimeMixer import TimeMixer
 from .diffusion_utils import GaussianDiffusion, make_beta_schedule, sinusoidal_embedding
 from .irregular import GRUD, mTAN, LatentODE, NeuralCDE, Raindrop
 
@@ -28,7 +29,7 @@ forecasting_models = [
     "PatchTST", "iTransformer",
     "TSMixer", "Crossformer", "SCINet", "TimesNet",
     "CATS", "FITS", "FreTS",
-    "SegRNN",
+    "SegRNN", "TimeMixer",
 ]
 
 generation_models = [